Basic functionality for engine perft

+++ include/past_move.h
@@ -16,7 +16,7 @@ struct PastMove {
     int promotion_piece;
     uint16_t enpassant;
     
-    PastMove(Move m, int pmi, int cp, int cs, uint8_t cr, int ep):
+    PastMove(Move m, int pmi, int cp, int cs, uint8_t cr, uint16_t ep):
         from(m.from), to(m.to), captured_piece(cp), piece_moved(pmi), captured_square(cs),
         castle_rights(cr), promotion_piece(m.promotion), enpassant(ep) {}
 };
